pyproject.toml
[project]
name = "odysseia-protect"
version = "0.1.0"
description = "A Discord bot for advanced file management."
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.8"
dependencies = [
    "SQLAlchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "discord.py>=2.3.2",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.4",
    "alembic>=1.13.0",
    'uvloop>=0.19.0; sys_platform != "win32"',
]

[project.optional-dependencies]
test = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]

[tool.uv]
# This section can be used for uv-specific configurations if needed in the future.
